feat: Add components.json for shadcn/ui configuration
This commit introduces the `components.json` file, which configures
`shadcn/ui` for the project.  It specifies the default style,
configures tailwind integration (including base color, CSS variables,
and prefix), defines aliases for common directories (components, utils,
lib, hooks, ui) and sets the icon library to "lucide". The 'rsc' is set to false and tsx to true. This file is essential for customizing the appearance and behaviour of UI components.