refactor: Use `<a>` tag for invoice download link in InvoicesIndex
Replaced `<Link>` component with `<a>` tag for the "Download Invoice" button in the `InvoicesIndex` component.
This change ensures the invoice is downloaded directly via a new tab, as intended by the `target="_blank"` attribute.
The `<Link>` component was interfering with the desired download behavior. Using a standard `<a>` tag provides the expected functionality of opening the route in a new tab/window which triggers a download.